lib/gstreamer: add common decoder base
Add a decoder base class for gstreamer to share
common code between msdk and vaapi elements.
